Fix routing in Next.js
