+++ frontend/src/pages/index.tsx
@@ -1,27 +1,41 @@
 import type {NextPage} from 'next'
-import Head from 'next/head'
-import MyApp from "./_app";
-import {StoreProvider} from "../stores";
+import {observer} from "mobx-react";
+import {useState} from "react";
+import {query, useQuery} from "../libs/dbConnection";
 
-const Home: NextPage = () => {
-  return (
-      <div className="">
-          <Head>
-              <title>YAMS</title>
-              <meta name="description"
-                    content="Yet Another Management System - Management for Clients, Animals and Bills"/>
+const Home: NextPage = observer(() => {
+    const [entryText, setEntryText] = useState("")
+    const [entries, refreshEntries] = useQuery("SELECT content FROM entry ORDER BY content", undefined, 1000)
+    const entriesExtracted = entries.length > 0 ? entries[0].result : []
 
-              <link rel="shortcut icon" href="/favicon.ico"/>
-              <link rel="icon" href="/favicon.ico"/>
-          </Head>
+    async function addEntry(text: string) {
+        console.log("Adding entry")
+        await query("CREATE entry SET content = $content", {
+            "content": text
+        })
 
-          <main className="">
-              <StoreProvider>
-                  <MyApp/>
-              </StoreProvider>
-          </main>
-      </div>
-  )
-}
+        refreshEntries()
+    }
+
+    return (
+        <main className="">
+            <div>
+                <div className="flex flex-row space-x-4">
+                    <p className="text-lg">Hello World!</p>
+                    <div className="flex flex-col space-y-2">
+                        <input onChange={(e) => setEntryText(e.target.value)}/>
+                        <button className="border rounded-lg bg-gray-400 text-black hover:bg-gray-200"
+                                onClick={(e) => addEntry(entryText)}>Add the Entry
+                        </button>
+                    </div>
+                    <div className="flex flex-col">
+                        {entriesExtracted.map((item: any, index: number) => <p
+                            key={index}>{item.content}</p>)}
+                    </div>
+                </div>
+            </div>
+        </main>
+    )
+})
 
 export default Home
